+++ src/angular-websocket.js
@@ -201,6 +201,9 @@ function $WebSocketProvider($rootScope, $q, $timeout, $websocketBackend) {
 
   $WebSocket.prototype._onCloseHandler = function _onCloseHandler(event) {
     var self = this;
+    if ((this.reconnectIfNotNormalClose && event.code !== this._normalCloseCode) || this._reconnectableStatusCodes.indexOf(event.code) > -1) {
+      this.reconnect(event);
+    }
     if (self.useApplyAsync) {
       self.scope.$applyAsync(function() {
         self.notifyCloseCallbacks(event);
@@ -208,9 +211,6 @@ function $WebSocketProvider($rootScope, $q, $timeout, $websocketBackend) {
     } else {
       self.notifyCloseCallbacks(event);
       self.safeDigest(true);
-    }
-    if ((this.reconnectIfNotNormalClose && event.code !== this._normalCloseCode) || this._reconnectableStatusCodes.indexOf(event.code) > -1) {
-      this.reconnect();
     }
   };
 
@@ -308,12 +308,15 @@ function $WebSocketProvider($rootScope, $q, $timeout, $websocketBackend) {
     return promise;
   };
 
-  $WebSocket.prototype.reconnect = function reconnect() {
+  $WebSocket.prototype.reconnect = function reconnect(event) {
     this.close();
 
     var backoffDelay = this._getBackoffDelay(++this._reconnectAttempts);
 
     var backoffDelaySeconds = backoffDelay / 1000;
+    if (typeof event === 'object') {
+      event.reconnectDelaySeconds = backoffDelaySeconds;
+    }
     if (this.consoleLogReconnect) {
       console.log('Reconnecting in ' + backoffDelaySeconds + ' seconds');
     }